+++ app/build_specs/formatter_selectors.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from app.build_specs.definitions import Selector, Mapper
+from app.definitions.mapper import Selector, Mapper
 from app.definitions.spec import BuildSpecError
 from app.formatters.formatter import Formatter
 from app.formatters.looping_formatter import LoopingFormatter
