+++ app/build_specs/mappings.py
@@ -1,24 +1,20 @@
 from sdx_gcp.errors import DataError
 
-from app.build_specs.definitions import Mapper
-from app.build_specs.reader import BuildSpecRepository
+from app.definitions.mapper import Mapper, SpecMapping
+from app.definitions.spec_repo import BuildSpecRepository
 from app.definitions.spec import BuildSpec
 from app.definitions.data import SurveyMetadata
 from app.build_specs.spec_selectors import Selector
 
 
-class SpecMapping[T](Mapper[T, str]):
-
-    def __init__(self, mappings: dict[str, Selector[T, str]], repository: BuildSpecRepository):
-        super().__init__(mappings)
-        self._repository = repository
+class BuildSpecMapping[T](SpecMapping[T]):
 
     def get_build_spec(self, survey_id: str, discriminator: T) -> BuildSpec:
         spec_name: str = self.get(survey_id, discriminator)
         return self._repository.get_build_spec(spec_name)
 
 
-class BuildSpecMapping(SpecMapping[SurveyMetadata]):
+class PckSpecMapping(BuildSpecMapping[SurveyMetadata]):
 
     def __init__(self, mappings: dict[str, Selector[SurveyMetadata, str]], repository: BuildSpecRepository):
         super().__init__(mappings, repository)
@@ -31,7 +27,7 @@ class BuildSpecMapping(SpecMapping[SurveyMetadata]):
         return selector.choose(survey_metadata)
 
 
-class PrepopBuildSpecMapping(SpecMapping[str]):
+class PrepopBuildSpecMapping(BuildSpecMapping[str]):
 
     def __init__(self, mappings: dict[str, Selector[str, str]], repository: BuildSpecRepository):
         super().__init__(mappings, repository)
