+++ app/config/formatters.py
@@ -1,7 +1,4 @@
 from app.build_specs.formatter_selectors import FormatterSelector, FormatterMapping
-from app.build_specs.mappings import BuildSpecMapping, PrepopBuildSpecMapping
-from app.build_specs.reader import BuildSpecFileRepository
-from app.build_specs.spec_selectors import BuildSpecSelector, BuildSpecPeriodSelector, PrepopSelector
 from app.formatters.cora_formatter import CORAFormatter, MESFormatter
 from app.formatters.cora_looping_formatter import CORALoopingFormatter
 from app.formatters.cs_formatter import CSFormatter
