+++ app/definitions/spec_repo.py
@@ -1,0 +1,7 @@
+from app.definitions.spec import BuildSpec
+
+
+class BuildSpecRepository:
+
+    def get_build_spec(self, spec_name: str) -> BuildSpec:
+        pass
